+++ js/src/markdown/processors.ts
@@ -7,6 +7,7 @@ import rehypeSanitize from "rehype-sanitize"
 import rehypeHighlight from "rehype-highlight"
 
 import { rehypeAccessibleSuggestions } from "./plugins/rehypeAccessibleSuggestions"
+import { rehypeSuggestionCards } from "./plugins/rehypeSuggestionCards"
 import { remarkEscapeHtml } from "./plugins/remarkEscapeHtml"
 import { rehypeExternalLinks } from "./plugins/rehypeExternalLinks"
 import { rehypeUncontrolledInputs } from "./plugins/rehypeUncontrolledInputs"
@@ -30,6 +31,7 @@ export const markdownProcessor = unified()
   .use(rehypeUnwrapBlockCEs)
   .use(rehypeUncontrolledInputs)
   .use(rehypeAccessibleSuggestions)
+  .use(rehypeSuggestionCards)
   .use(rehypeExternalLinks)
   .use(rehypeHighlight, { detect: false, ignoreMissing: true })
   .freeze()
@@ -42,6 +44,7 @@ export const markdownProcessor = unified()
 export const htmlProcessor = unified()
   .use(rehypeUncontrolledInputs)
   .use(rehypeAccessibleSuggestions)
+  .use(rehypeSuggestionCards)
   .use(rehypeExternalLinks)
   .freeze()